Rename module to be less cursed
