+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools >= 61.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "ttmpRL"
+name = "chessrl"
 version = "0.0.1"
 dependencies = [
   "cairosvg==2.8.2",
